Update AEAD limits for larger packets
The analysis describes conditions where higher limits might be allowed,
and provides a formula in each case that can be used to determine other
limits more readily.

The number of changes I've made here are somewhat frightening.  This is
why I resisted doing this.  This will need careful double-checking.

The actual numbers don't concern me much, it's the changes to the
structure that need to be looked at.  Some of the changes are editorial,
but they still need extra eyes.

Closes #3701.