Prohibit using old keys for sending
In this context, "newer" was confusing as the code permits the creation
of the next set of packet protection keys when switching over.  As the
goal of the text was to ensure that old keys not be used, using that
formulation avoids any confusion that might be caused from not having a
reference against which to decide which is "newer".

Closes #4199.