DCID is 0..20 and clarify...
... that after Retry the use of the new Destination Connection ID does
not provide address validation (this was previously obtusely stated).

Note that this critically says that a Destination Connection ID of any
length is not sufficient to prove that the server got the packet.
This is because the client doesn't know how hard the value is to guess,
even if the value is 8 byte or more long and might appear to contain
sufficient entropy (it could be a deterministic function of the client
address, for example).