Pad when validating addresses
This isn't obvious, but if you are validating an address, you are also
validating the path to that address.  If you don't pad these packets,
then you don't know that the path supports a reasonable PMTU.

Require padding of datagrams that carry packets that carry
PATH_CHALLENGE.  This neatly solves the problem.

Closes #4216.