Restructure the abstract and overview
This takes some of the concepts that @MikeBishop added in #4326 and
massages those into the introduction.  The net effect is two-fold: a
little more substance in this section and the addition of prominent,
early references to the key supporting documents.

It removes the bullet feature list and puts that into the abstract as a
single sentence.

I have also reordered this slightly.  There is no perfect ordering here,
but this no longer very closely matches document structure.