Stop using "may"
In most cases, this turned out to be the wrong word anyway.

There is "can", which is a simple statement of fact.  There is "could",
for stating that something is possible.  There is "MAY", once, which
might be wrong.  There is also cases where the word wasn't needed at
all.

I know that "may" is used in all of these cases in the strictly
colloquial sense, with no normative intent, but it is a tiny bit
distracting.