invariants: clarify authentication requirements for version negotiation
There is no need to authenticate specificaly the bits contained in the
Version Negotiation packet, just the information that is being acted
on.  Since the authentication of this information is likely to occur
out of band from the Version Negotiation packet itself (e.g., by
including the list of supported versions in a protocol element in a
subsequent connection and/or handshake), attempt to clarify that
only the information needs to be authenticated, not the packet itself.