Closure alerts are errors too
This changes the shape of the requirement from one that proscribes the
behaviour of a TLS stack to the reaction of QUIC to a warning level
alert.  This recommends treating them as fatal as QUIC never
deliberately closes TLS.

This is a normative change that should have no practical effect on
implementations.

Closes #4485.