Future-proof how we cover Version Negotiation
Attempt to more accurately reflect the expectations of the WG for
how version negotiation will be defined and what endpoint behavior
should be in the absence of a version negotiation mechanism.
In particular, this work is to be done by the IETF, not by arbitrary
future versions of QUIC, so we retain control over how it will be
developed and deployed.

Also accommodate the possibility of new versions of QUIC appearing
before version negotiation is defined.

Fixes: #4607