Reason phrase might be incomprehensible
Though we recommend the use of text in UTF-8, we don't provide all of
the machinery necessary for true internationalization, like multiple
translations and all that mess.  Be very explicit about that.

(This is a recurring problem; we had the same problem with HTTP/2 and it
might be worth updating BCP 18 for this.)

Closes #4647.