Finesse relationship of resumption and 0-RTT
Resumption is not quite the basis of 0-RTT, since it is also allowed to
pre-provision an external PSK and other parameters and do 0-RTT that
way.  This also allows a rewording to avoid implying that 0-RTT would
need to be enabled by configuration, since it is supposed to just be
"on by default".

Fixes: #4477