Two small tweaks to connection IDs
The value the client puts in the Destination Connection ID field is not
random, but unpredictable.  But we could be clearer and just say what
this refers to.

The second error here was that the server can't use the value it puts in
Retry.  Rather than talk about retaining the values, I'm striking that
clause.

Closes #4677.