Tidy the XML
The purpose of this tool is to ensure that we can produce an XML file
that won't diverge too much from what the RPC produces.

The main thing this does is add \<bcp14\> elements as necessary.  But it
also tweaks indentation and removes XML garbage.  More can be added as
we discover the need.