Get case of Stateless Reset right
Stateless Reset is capitalized when referring to the thing that is sent,
but stateless reset as a concept is lowercase.

This is to be consistent with packet naming, even if this isn't a
packet. This also removes "stateless reset packet" and "Stateless Reset
packet" and variations of the same from the document by removing
"packet" in all cases.  A stateless reset is a UDP datagram always.